bin: iot-rest-api-server script added
Instead of running index.js when this package is installed
with the `-g` flag to npm it will install the script called
iot-rest-api-server which runs the server.
