+++ package.json
@@ -13,6 +13,9 @@
   "dependencies": {
     "iotivity-node": "^1.3.0-0"
   },
+  "bin": {
+    "iot-rest-api-server": "bin/iot-rest-api-server.js"
+  },
   "devDependencies": {
     "gulp": "^3.8.11",
     "gulp-nodemon": "^2.0.3"
